refactoring analysis methods
