+++ api/app/resources/requests.py
@@ -6,21 +6,18 @@ from flask import request, jsonify, g, current_app
 from flask_restplus import Namespace, Resource, fields, cors
 from sqlalchemy.orm.exc import NoResultFound
 from sqlalchemy.exc import SQLAlchemyError
-from sqlalchemy import func, desc, asc, text, exc
+from sqlalchemy import func, text, exc
 from sqlalchemy.inspection import inspect
-import types
-import re
-import sys
 
-from marshmallow import ValidationError
 from app import oidc
 from app.auth_services import required_scope, AuthError
-from app.models import db, User, State, Name, NameSchema
+from app.models import db, User, State, NameSchema
 from app.models import Request as RequestDAO, RequestsSchema
 from app.models import DecisionReason
 
 from app.utils.util import cors_preflight
-from .solr import SolrQueries
+from app.analytics.solr import SolrQueries
+from app.analytics.restricted_words import get_restricted_words_conditions
 
 # Register a local namespace for the requests
 api = Namespace('nameRequests', description='Name Request System - Core API for reviewing a Name Request')
@@ -436,98 +433,8 @@ class RequestsAnalysis(Resource):
             return jsonify(conflicts), 200
 
         else:
-            return RequestsAnalysis.get_restricted_words_conditions(nrd_name.name)
-
-    @staticmethod
-    def get_restricted_words_conditions(corp_name):
-        """ 1. put all possible restricted words/phrases in a list
-                          - used later to compare against sql fn
-                       parse corp_name from snake_case into sql format
-        """
-        word_list = corp_name.split()
-
-        # this adds in all possible phrases that are two or more words to word_list
-        phrases = []
-        phrase = ''
-        for indx, word in enumerate(word_list):
-            for possible_phrase in word_list[indx:]:
-                if phrase != '':
-                    phrase += possible_phrase
-                    phrases.append(phrase)
-                else:
-                    phrase += possible_phrase
-                if indx < len(word_list):
-                    phrase += ' '
-            phrase = ''
-
-        word_list = word_list + phrases
-        """------------------------------------------------------"""
-
-        """ 2. get words/phrases in corp_name that are restricted
-                - query for list of all restricted words
-                    - compare these words to word_list
-        """
-
-        get_all_restricted_words_sql = text("select * from restricted_word;")
-        try:
-            restricted_words_obj = db.engine.execute(get_all_restricted_words_sql)
-
-        except exc.SQLAlchemyError:
-            print(exc.SQLAlchemyError)
-            return jsonify({"message": "An error occurred accessing the restricted words."}), 500
-        except AttributeError:
-            return jsonify({"message": "Could not find any restricted words."}), 404
-        restricted_words_dict = []
-        for row in restricted_words_obj:
-            for word in word_list:
-                if row[1] == word:
-                    restricted_words_dict.append({'id':row[0],'phrase':row[1]})
-        """-----------------------------------------------------------------"""
-
-        """ 3. get condition info based on word_id for each restricted word """
-
-        restricted_words_conditions = []
-        for word in restricted_words_dict:
-            get_cnd_id_sql = text("select cnd_id from restricted_word_condition where word_id = {}".format(word['id']))
-            try:
-                cnd_id_obj = db.engine.execute(get_cnd_id_sql)
-                cnd_ids = cnd_id_obj.fetchall()
-
-                cnd_obj_list = []
-                for id in cnd_ids:
-                    cnd_id = id[0]
-                    get_cnd_sql = text("select * from restricted_condition where cnd_id = {}".format(cnd_id))
-                    cnd_obj_list.append(db.engine.execute(get_cnd_sql))
-
-                cnd_info = []
-                for obj in cnd_obj_list:
-                    obj_tuple = obj.fetchall()[0]
-                    cnd_text = obj_tuple[1]
-                    cnd_allow_use = obj_tuple[2]
-                    cnd_consent_req = obj_tuple[3]
-                    cnd_consent_body = obj_tuple[4]
-                    cnd_instr = obj_tuple[5]
-
-                    cnd_info.append({'id': cnd_id,
-                                    'text': cnd_text,
-                                    'allow_use': cnd_allow_use,
-                                    'consent_required': cnd_consent_req,
-                                    'consenting_body': cnd_consent_body,
-                                    'instructions': cnd_instr})
-                restricted_words_conditions.append({'word_info': word, 'cnd_info': cnd_info})
-            except exc.SQLAlchemyError:
-                print(exc.SQLAlchemyError)
-                return jsonify(
-                    {"message": "An error occurred accessing the condition for {}.".format(word['id'])}), 500
-            except AttributeError:
-                return jsonify({"message": "Could not find any condition info for {}.".format(word['id'])}), 404
-            except:
-                # print('error')
-                cnd_info = 'Not Available'
-                restricted_words_conditions.append({'word_info': word, 'cnd_info': cnd_info})
-        """------------------------------------------------------------------------------------"""
-
-        return jsonify({"restricted_words_conditions": restricted_words_conditions}), 200
+            restricted, code = get_restricted_words_conditions(nrd_name.name)
+            return jsonify(restricted), code
 
 
 @cors_preflight("GET, PUT, PATCH")
